refactor: take away localStorage logic
